18MS: Move to new structure
Part of #3849
