+++ lib/engine/game/g_18_ms/step/lay_tile_with_chattanooga_check.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Engine
-  module Step
+  module Game
     module G18MS
       module LayTileWithChattanoogaCheck
         ACTION_HEX_OF_INTEREST = 'B10'
