+++ lib/engine/game/g_18_ms/step/track.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/track'
+require_relative 'lay_tile_with_chattanooga_check'
+
+module Engine
+  module Game
+    module G18MS
+      module Step
+        class Track < Engine::Step::Track
+          include LayTileWithChattanoogaCheck
+        end
+      end
+    end
+  end
+end
